+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `status` command for getting current endpoint and/or context.
 
+### Changed
+
+- Uses `default` context in case none is specified.
+
 ### Fixed
 - `variable` commands would always use the default context instead of the active one, causing completly broken behavior.